dev-requirements.txt
-r requirements.txt
django-debug-toolbar==1.6
django-userswitch
